+++ src/index.ts
@@ -8,12 +8,8 @@ import { context } from "~/context";
 import { env } from "~/env";
 import { schema } from "~/schema";
 import { handler as razorpayCapture } from "~/webhook/capture";
-import { uploadRouter } from "./uploadthing";
-import jwt, { JwtPayload } from "jsonwebtoken";
-import { secrets } from "./utils/auth/jwt";
-import { authenticateUser } from "./utils/auth/authenticateUser";
-import { PrismaClient } from "@prisma/client";
-import { prisma } from "./utils/db/prisma";
+import { uploadRouter } from "./uploadthing/FileRouter";
+import authMiddleware from "./uploadthing/middleware";
 
 const yoga = createYoga({
   context,
@@ -22,40 +18,6 @@ const yoga = createYoga({
 });
 
 const app = express();
-const authMiddleware = async (
-  req: express.Request,
-  res: express.Response,
-  next: express.NextFunction,
-) => {
-  const authHeader = req.headers.authorization;
-  const token = authHeader?.split(" ")[1];
-
-  if (!token) {
-    return res.status(401).json({ error: "Unauthorized: No token provided" });
-  }
-
-  try {
-    const tokenPayload = jwt.verify(
-      token,
-      secrets.JWT_ACCESS_SECRET as string,
-    ) as JwtPayload;
-
-    if (tokenPayload.userId) {
-      const user = await prisma.user.findUnique({
-        where: { id: tokenPayload.userId },
-      });
-
-      if (!user) {
-        return res.status(401).json({ error: "Unauthorized: User not found" });
-      }
-    }
-
-    next(); // Call next to proceed to the next middleware or route handler
-  } catch (e) {
-    console.error("Token verification error:", e);
-    return res.status(401).json({ error: "Unauthorized: Invalid token" });
-  }
-};
 app.use(
   cors({
     origin: env.FRONTEND_URL,
@@ -83,19 +45,3 @@ app.use(
 app.listen(env.PORT, () => {
   console.log(`🚀 Server ready at: http://localhost:4000/graphql`);
 });
-
-// const { upload } = config;
-// const { upload: easterUpload } = easterConfig;
-// const { upload: idUpload } = idUploadConfig;
-
-// app.post("/cloudinary/upload/:eventName", upload.single("image"), imageUpload);
-// app.post("/easter-egg/upload", easterUpload.single("image"), imageUpload);
-// app.post("/id/upload", idUpload.single("image"), imageUpload);
-
-// import { uploader as imageUpload } from "./cloudinary/upload";
-// import { config } from "./cloudinary/config";
-// import { config as easterConfig } from "./cloudinary/easterConfig";
-// import { config as idUploadConfig } from "./cloudinary/idUpload";
-
-// // import "./certificate.ts";
-// import jwt from 'jsonwebtoken';
